refactor recording form into intent-driven full-stack resource
